Add basic cast support (#64)
This introduces Basic cast support for Flutter. The API is align as much as possible with the [React Native SDK](https://github.com/bitmovin/bitmovin-player-react-native/pull/262).

## New API
 - `BitmovinCastManager`
 - `player` cast methods (`castVideo`, `castStop`, `isCasting`...)

### Not yet supported
 - Cast events

## Issues encountered
 - The Cast SDK expects an AppCompat `Activity` (`FragmentActivity`), which the recommended `FlutterActivity` is not.
     - `MainActivity` now extends `FlutterFragmentActivity`
 - The Cast SDK expects an AppCompat style
     - The `Theme.AppCompat.NoActionBar` is now used instead of `LaunchApp`
 - `com.bitmovin.player.casting.ExpandedControllerActivity` was not accessible in the sample app
    - Made the player an `api` dependency

## Design
### `BitmovinPlayerManager`
Flutter's `Player` is a plain object with an async initialization. Aka, the Dart Player object is created immediately, but it's resources (native player, channels...) are created in the background. The client doesn't know when it is done [^1].
This was chosen for convenience as Dart has no async constructor.

The drawback of this approach is that the client code doesn't control the initialization sequence of the object explitly.

Given that the `BitmovinPlayerManager` must be created before the `Player`, the same patter is not applicable. Instead the `BitmovinPlayerManager` singleton is accessible through an async static `create` factory. This ensure that the client can initialize it reliably before creating a `Player`.

Nevertheless as the player creation is not async, mixing the async and sync code is not very convenient and leads to the cast use case needing a more complex setup than the non cast use-case. This is reflected in the complexity of the cast sample included in this PR.

One possible solution to avoid this would be to implicitly create the cast manager on the platform side if the player config has cast enabled set to true. Nevertheless this is no inlined with our other SDKs, which need explicit initialization of the `BitmovinCastManager`. As such this alternative has not been implemented, but **feedback on this proposal is welcomed**.

[^1]: The async part is implicitly `await`ed when an async method of the player is called. This is an implementation detail that should not be relied on by the client.

### `BitmovinPlayerManager.updateContext`
`updateContext` is supposed to be called for any new activity. Nevertheless the Flutter native plugin already gets a callback on Activity attach and detach. This PR has `updateContext` called automatically if the `BitmovinPlayerManager` is initialized.

As such there should be no need to call `updateContext` from the Dart side. It is still exposed for alignment with React Native and because I didn't full understand the Flutter plugin Activity lifecycle when I started this PR.

Thus the `updateContext` is not exposed in the Dart API.